// src/frontend/src/api/identity.d.ts
export interface components {
    schemas: {
        ChangePasswordRequest: {
            currentPassword: string;
            newPassword: string;
        };
        HttpValidationProblemDetails: {
            type?: null | string;
            title?: null | string;
            /** Format: int32 */
            status?: null | number;
            detail?: null | string;
            instance?: null | string;
            errors?: {
                [key: string]: string[];
            };
        };
        ProblemDetails: {
            type?: null | string;
            title?: null | string;
            /** Format: int32 */
            status?: null | number;
            detail?: null | string;
            instance?: null | string;
        };
        ProfileResponse: {
            /** Format: uuid */
            id: string;
            email: string;
            displayName: string;
            roles: string[];
            permissions: string[];
        };
        RecoveryCompleteRequest: {
            token: string;
            newPassword: string;
        };
        RecoveryRequest: {
            email: string;
        };
        RegisterRequest: {
            email: string;
            password: string;
            displayName: string;
        };
        SignInRequest: {
            email: string;
            password: string;
        };
        TokenRequest: {
            token: string;
        };
        TokenResponse: {
            accessToken: string;
            tokenType: string;
            /** Format: int32 */
            expiresIn: number;
        };
        UpdateProfileRequest: {
            displayName: string;
        };
        UserPage: {
            items: components["schemas"]["UserSummary"][];
            /** Format: uuid */
            nextCursor: null | string;
        };
        UserSummary: {
            /** Format: uuid */
            id: string;
            email: string;
            displayName: string;
            disabled: boolean;
            emailConfirmed: boolean;
            /** Format: date-time */
            createdAt: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}